+++ packages/help-panel/src/components/HelpPanel/HelpPanel.tsx
@@ -5,10 +5,8 @@
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-import { Color } from '@harness/design-system'
 import React from 'react'
-import { Error, useContentful } from '../../HelpPanelContext'
-import { ContentType, IHelpPanel } from '../../types/contentfulTypes'
+import { Error, HelpPanelContext } from '../../HelpPanelContext'
 import DefaultContainer from './Containers/DefaultContainer/DefaultContainer'
 import FloatingContainer from './Containers/FloatingContainer/FloatingContainer'
 import HelpPanelContent from './HelpPanelContent/HelpPanelContent'
@@ -28,17 +26,7 @@ const HelpPanel: React.FC<HelpPanelProps> = props => {
   const { referenceId, type } = props
   let floatingBtnRef: HTMLButtonElement
 
-  const {
-    data = {
-      backgroundColor: Color.WHITE,
-      articles: []
-    },
-    loading,
-    error
-  } = useContentful<IHelpPanel>({
-    referenceId,
-    content_type: ContentType.helpPanel
-  })
+  const { error } = React.useContext(HelpPanelContext)
 
   if (error === Error.ERROR_INITIALIZING_CONTENTFUL) {
     return null
@@ -52,8 +40,7 @@ const HelpPanel: React.FC<HelpPanelProps> = props => {
             floatingBtnRef = node
           }}>
           <HelpPanelContent
-            data={data}
-            isLoading={loading}
+            referenceId={referenceId}
             onClose={() => {
               floatingBtnRef?.click()
             }}
@@ -61,11 +48,11 @@ const HelpPanel: React.FC<HelpPanelProps> = props => {
         </FloatingContainer>
       )
     case HelpPanelType.CONTENT_ONLY:
-      return <HelpPanelContent data={data} isLoading={loading} onClose={props.onClose} />
+      return <HelpPanelContent referenceId={referenceId} onClose={props.onClose} />
     default:
       return (
         <DefaultContainer>
-          <HelpPanelContent data={data} isLoading={loading} />
+          <HelpPanelContent referenceId={referenceId} />
         </DefaultContainer>
       )
   }