+++ packages/help-panel/src/components/HelpPanel/HelpPanelContent/HelpPanelContent.tsx
@@ -10,34 +10,45 @@ import { Color } from '@harness/design-system'
 import Container from '../../Container'
 import Header from '../Header/Header'
 import Footer from '../Footer/Footer'
-import { IHelpPanel } from '../../../types/contentfulTypes'
+import { IHelpPanel, ContentType } from '../../../types/contentfulTypes'
 import css from './HelpPanelContent.module.css'
 import Skeleton from '../../Skeleton/Skeleton'
 import Article from '../../Article/Article'
+import { useContentful } from '../../../HelpPanelContext'
+import ErrorScreen from '../ErrorScreen/ErrorScreen'
 
 interface Props {
-  data: IHelpPanel
-  isLoading: boolean
+  referenceId: string
   onClose?: () => void
 }
 
 export const HEADER_FOOTER_HEIGHT = 64
-const HelpPanelContent: React.FC<Props> = ({ data, onClose, isLoading }) => {
-  const { backgroundColor = Color.BLUE_50, title, articles } = data || {}
+const HelpPanelContent: React.FC<Props> = ({ onClose, referenceId }) => {
+  const { data, loading, error } = useContentful<IHelpPanel>({
+    referenceId,
+    content_type: ContentType.helpPanel
+  })
+
+  const { articles, title, backgroundColor = Color.BLUE_50 } = data || {}
 
   const renderContent = () => {
     return (
       <>
         <Header title={title} onClose={onClose} />
-        <Container
-          className={css.centerContainer}
-          padding={{ left: 'xlarge', right: 'xlarge' }}
-          width="100%"
-          style={{ top: HEADER_FOOTER_HEIGHT, height: `calc(100% - ${HEADER_FOOTER_HEIGHT * 2}px)` }}>
-          {articles?.map(article => (
-            <Article key={article.sys.id} {...article.fields} />
-          ))}
-        </Container>
+        {!error && articles?.length !== 0 ? (
+          <Container
+            className={css.centerContainer}
+            padding={{ left: 'xlarge', right: 'xlarge' }}
+            width="100%"
+            style={{ top: HEADER_FOOTER_HEIGHT, height: `calc(100% - ${HEADER_FOOTER_HEIGHT * 2}px)` }}>
+            {articles?.map(article => (
+              <Article key={article.sys.id} {...article.fields} />
+            ))}
+          </Container>
+        ) : (
+          <ErrorScreen />
+        )}
+
         <Footer />
       </>
     )
@@ -45,8 +56,8 @@ const HelpPanelContent: React.FC<Props> = ({ data, onClose, isLoading }) => {
 
   return (
     <Container background={backgroundColor} style={{ height: '100%' }}>
-      {isLoading ? (
-        <Container padding={{ top: 'xlarge', bottom: 'xlarge' }}>
+      {loading ? (
+        <Container padding="xlarge">
           <Skeleton />
         </Container>
       ) : (