+++ src/index.ts
@@ -137,6 +137,7 @@ export {
   MultiTypeInputType,
   RUNTIME_INPUT_VALUE,
   EXPRESSION_INPUT_PLACEHOLDER,
+  EXECUTION_TIME_INPUT_VALUE,
   MultiTypeIcon,
   MultiTypeIconSize
 } from './components/MultiTypeInput/MultiTypeInputUtils'